exception handler

+++ src/main/java/todolistApp/hundredSteps/controller/TodoController.java
@@ -93,6 +93,10 @@ public class TodoController {
 	
 	@RequestMapping(value="/delete-todo", method = RequestMethod.GET)
 	public String deleteToDo(@RequestParam int id) {
+		
+		if(id==1)
+			throw new RuntimeException("Something went wrong");
+		
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
